refactor(usecase): use unique layer for order usecase
