+++ internal/core/usecase/delete_product.go
