+++ internal/core/usecase/healthcheck_ping.go
