+++ internal/core/usecase/mappers/product.go
